+++ db/updates/update_2.4-2.5.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `ingredients` CHANGE `usage_type` `usage_type` VARCHAR(255) CHARACTER SET utf8 COLLATE utf8_bin NULL;